Disregard the 'mounted' dataset property.
If `/etc/mtab` is not a symlink to `/proc/mounts`, and if it contains stale
records, then `zfs list` returns incorrect 'mounted' values that cause
corresponding datasets to be ignored.

The mounted check is unnecessary (and the problem), since mountall checks if
the mountpoints are already mounted before it spawns a mount job - but zfs list
checks mtab before it's been written by mountall (and any stale mountpoints get
misrepresented). Removing that check, everything seems to work fine (mounts
correctly on startup, doesn't overmount things when run from the command line
in a running session).

Amended-by: Darik Horn <dajhorn@vanadac.com>
Closes: dajhorn/pkg-zfs#70