Import mountall-2.42ubuntu0.3 for quantal.
The upstream lp:ubuntu/quantal/mountall repository is stale and does not have a
branch or tag for the mountall-2.42ubuntu0.3 release that is currently
published in precise-updates.

NB: https://bugs.launchpad.net/ubuntu/+source/mountall/+bug/1103319

This bug report was closed as invalid because Ubuntu policy does not
technically require a source code control for system updates.