Add debian/source/format file.
And set "3.0 (quilt)" so that the patch series is automatically applied by the
Launchpad builder.

This requires the deletion of `.bzr-builddeb`, which is in the upstream
Launchpad repository but not in the official distro tarball. Otherwise:

dpkg-source: info: local changes detected, the modified files are:
 mountall/.bzr-builddeb/default.conf
dpkg-source: error: aborting due to unexpected upstream changes