Handle duplicate mountpoints.
When mountall is iterating in try_mounts() on new mount point, change
is_parent() to return false when a record is tested against itself
because a mount point cannot be a root of itself.

Mountall does not preserve the behavior of `mount -a` if more than
one device shares a mountpoint in the /etc/fstab file.  Suppose:

	/var/tmp/alfa   /mnt/charlie  ext2  loop  0  0
	/var/tmp/bravo  /mnt/charlie  ext2  loop  0  0

Both filesystems are mounted on /mnt/charlie by `mount -a`, but
bravo is ignored by `mountall`. This seems to be an artifact of the fix for
LP ##443035 in zfsonlinux/mountall@eca315d06ae4a2913f9b2ec994c68c45fead912f.

Furthermore, mountall hangs if such mounts are ZFS datasets or other
virtual filesystems that are not backed by a real device node.

Closes: zfsonlinux/mountall#5